cdfdif(RTmodels)	R Documentation

Compute simultaneous probability of (a) response time(s) and error under Ratcliff's diffusion model.

Description

CDFDif Returns the probability Prob(RT <= t, X=x), where RT is a single response time, and X is an indicator (correct or incorrect). cdfdif does the same, except that t is allowed to be an array.

Usage


CDFDif(t,x,nu,z,a,Ter,eta=0,sz=0,st=0);
cdfdif(t,x,nu,z,a,Ter,eta=0,sz=0,st=0)

Arguments

t 	 degrees of freedom parameter

x 	 TRUE (=correct) or FALSE (=incorrect)

a 	 the boundary separation

Ter 	 the mean non-decisional component time

eta 	 the standard deviation of the normal drift distribution

z 	 the mean starting point

sz 	 the spread of the starting point distribution

st 	 the spread of the non-decisional component time distribution

nu 	 the mean drift rate

Value

The probability/probabilities Prob(RT <= t, X=x).

Remark

Original Fortran sources due to Francis Tuerlinckx. C translation by Raoul Grasman.

Author(s)

Raoul Grasman rgrasman@uva.nl

References

Tuerlinckx, (2004), The efficient computation of the cumulative distribution and probability density functions in the diffusion model, Behavior Research Methods, Instruments, & Computers, vol. 36 (4): 702-716

Examples


##
cdfdif(c(0.3,0.5),1,nu=.2,z=.11,a=.2,Ter=.3,eta=.1,sz=0.05,st=0.1)

[Package RTmodels version 0.1 Index]