difsim.old(RTmodels)	R Documentation

Function to simulate diffusion processes model of Rattcliff for perceptual decisions

Description

Calling difsim results in the reaction times of both correct and incorrect responses of N trials, as obtained from a diffusion process. All parameters of the diffusion process can be modified. The underlying process is given by the stochastic differential equation dx = v dt + sqrt{var.v} * dW(t).

Usage


difsim.old(v, a, N, z = 0.5 * a, s = 0.1, var.v = 0, range.z = 0, Ter = 0, range.Ter = 0, stepsize = 0.0)

Arguments

v 	 Drift rate of the diffusion

a 	 Upper decision boundary. Lower decision boundary is 0.

N 	 The number of trials.

z 	 Starting point of the difusion, defaults to the center of the lower and upper boundaries.~~

s 	 Variance of dW(t)

var.v 	 Variance of the driftrate. Defaults to zero.

range.z 	 Radius of uniform distributed starting point around z.

Ter 	 Fixed constant for motor respons.

range.Ter 	 Radius of uniform distribution around Ter.

stepsize 	 Size (msec) of the steps taken in the random walk. Default value is appropriate voor common parameter values.

Details

For details on the meaning of the parameters, see the references.

Value

t.top 	 reaction times in trials hitting the boundary a ("correct" responses).

t.bot 	 reaction times in trials hitting the boundary 0 ("incorrect" responses).

Warning

This function is obsolete.

Author(s)

Raoul Grasman Eric-Jan Wagenmakers

References

Ratcliff & Tuerlinckx, (2002), Estimating parameters of the diffusion model: Approaches to dealing with contaminant reaction times and parameter variability, Psychonomic Bulletin & Review, vol. 9 (3): 438-481

Examples


  rt = difsim(v=.2,a=.2,z=.09,Ter=.250,N=1000);
  quantile(rt[[1]],seq(0.05,0.95,l=10))

[Package RTmodels version 0.1 Index]